Place a throwaway directory in /opt/lmi/, not in /tmp/
See:
  https://lists.nongnu.org/archive/html/lmi/2019-05/msg00029.html
To a native msw program running under cygwin, '/tmp' is something like
  D:\tmp
whereas to cygwin itself, '/tmp' is something like
  D:\cygwin64-lmi\tmp
. To avoid that unwanted contrariety, placed the throwaway directory
under /opt/lmi/ so that the cygwin "identity mount" maps both to the
same place.

Full paths seem preferable because of their unambiguity. For example, if
  $PERFORM /opt/lmi/bin/lmi_cli_shared --file="$throwaway_dir"/X.ill
is echoed into a nychthemeral log as
  wine /opt/lmi/bin/lmi_cli_shared --file=/opt/lmi/amd64/logs/tmp/X.ill
and that command fails, then it can be copied and pasted into a terminal
to be run manually without adjusting the terminal's $PWD.