Derive cover page from standard_page, not from numbered_page
Deriving a leaf class from a more-derived class is generally preferable
because it allows more behavior to be inherited. In this case, it was no
longer useful to override render() (because the decorative rectangle had
been removed), and it is better not to override get_extra_pages_needed()
(because blithely returning zero would silently become wrong if the
cover page ever needed more than one physical page).