Fix defect introduced 20190420T2117: non-minimal $PATH
Prior to commit 7ea51651e7a4, this block of (pseudo)code was correct:
  # a minimal $PATH had already been set long before
  {
  run commands that require only a minimal $PATH
  add lmi runtime paths to minimal $PATH
  run commands that require lmi runtime paths
  }
Commit 7ea51651e7a4 wrapped that block in a for-loop, rendering it
incorrect: $PATH was minimal, as intended, only for the first iteration
of the loop.

Anyone who doesn't believe that $PATH should contain nothing more than
is required also might not believe this matters. But proof is a higher
standard than belief.