Add git instructions to copy a branch into a master subdirectory
Motivation: To delete the odd/source_script_in_makefile branch, yet
retain its history in the repository. Given the circumstances, it
actually was possible to "replay" the history into the permanent gwc/
subdirectory. Deleting the files there will not remove their history.

It would have been better, of course, to develop this demonstration
in gwc/ rather than in a branch.