Fix defect committed 20190518T1712Z: script might not be in $PWD
Sourcing a script thus (commit 210257aee9f8 of 20190518T1712Z):
  . ./set_toolchain.sh
assumes that it resides in $PWD. Arguably that line should have been
  . $(srcdir)/set_toolchain.sh
but contemplating that reveals another defect:
  $(srcdir)/GNUmakefile
is intended to allow this makefile to be made from a different
directory, but $(srcdir) is not yet defined.

Changed these two lines to use $(CURDIR), which 'make' predefines.
That's the smallest change required for

  $make -C /opt/lmi/src/lmi -f /opt/lmi/src/lmi/GNUmakefile \
    check_concinnity

to work, and it's also the righteous change. However, this test case

  $cd /tmp
  $/opt/lmi/src/lmi/hooks/pre-commit

still fails, because commit 0b55f4657b6d of 20120301T1226 was
suboptimal; see the next commit.