Purify sourcing demonstration by removing a desirable refinement
Moved commands to set 'make' variables out of the shell script and back
into the makefile.

The parent commit is preferable for actual use. See:
  https://lists.nongnu.org/archive/html/lmi/2019-05/msg00052.html
et seq.

Ask the internet how to source a shell script in a makefile, and it will
answer that it simply can't be done--that the script must be changed.
This revision proves that it can be done, without changing the script.
That is this demonstration's purpose: to show how it can be done,
regardless of whether it should be done in exactly this way.