Fix defect introduced 20190426T2304Z: wrong variable used
Commit 36191bc99d of 20190426T2304Z intended to replace the literal
"/opt/lmi/local/bin" with an equivalent variable. The intention was
proper, but the wrong variable, $(bindir), was used. Fixed: now the
right variable, $(localbindir), is used instead.

The correct substitution was made here:
  20190326T2301Z commit dc3854e70d
and it survived the reversion here:
  20190426T1734Z commit 2d86f7e3ac
which happened to undo the substitution and restore the original
literal. The first literal-substitution changeset was disciplined
(vim ':%s' commands) because it had to be right. The single
substitution in commit 36191bc99d wasn't disciplined because it
seemed that it couldn't possibly be wrong. The original commands
were still in vim history, so typing over the literal was unwise,
especially when tidying up after a reversion.

This commit fixes this reported problem:
  https://lists.nongnu.org/archive/html/lmi/2019-05/msg00059.html
It's not yet ideal: files in $(compiler_runtime_files) are copied in
a different way (to a different subdirectory) in 'GNUmakefile'; and
running `make -n fardel` through shellcheck elicits:
  SC2157: Argument to -z is always false due to literal strings.
for the controlling condition here:
  [ -z "$(compiler_runtime_files)" ]
because of a trailing blank.