Generally prefer printf to echo
Removed a comment that incorrectly said shell builtins mustn't be used
in a script that is to be sourced. That was an improper generalization
of the actual truth that a zsh-specific builtin like 'print' mustn't be
used in a script that is to be sourced by dash; OTOH, POSIX builtins
like 'printf' are okay.