Add 'test_coding_rules' to the list of default targets
'hooks/pre-commit' fails if the 'test_coding_rules' binary is out of
date. That doesn't happen if 'make check_concinnity' is run before every
commit, but doing so optional (the hook does it anyway). However, it is
reasonable to expect that either 'make install' or 'make all' has been
run before any commit, to ensure that the changes compile; those targets
now ensure that this binary is up to date.

Reverted commit b30b5db803, which has been rendered unnecessary.

Alternative not chosen: adding this line
  make --directory=/opt/lmi/src/lmi custom_tools 2>&1 >/dev/null
to 'hooks/pre-commit' would ensure that the required binary has always
been built before it's needed. However, that hook already takes about
five and a half seconds, and it would take seven seconds with such a
change. The extra 'make' invocation seems too costly for a git hook.
The chosen solution has a lower incremental cost (adding a dependency
doesn't require reentering the top-level makefile), and 'make all' need
not be as fast as a hook script.