Use a minimum-age database for product verification
In order to validate 7PP, it will be necessary to retrieve the vector
quantity DB_NaarDiscount by querying the database with the minimum issue
age, so that the vector has one element for every possible attained age.

Alternatively, class product_database could be given an extra ctor that
would accomplish the same purpose. However, that seems like too big a
change for this special case, and the code committed is simple enough.