Split selected tables out of class BasicValues into a new TU
Added the new TU in this commit; deduplication will follow in the next.

For now at least, the new TU encompasses only 7702 and 7702A tables,
along with target-premium tables. The proximate goal is to provide
known-valid values on the 7702 basis:
  qx from built-in mortality tables of trusted provenance, and
  7Px, Ax, and 1/Ax from first principles,
as reliable touchstones for validating tables provided by pricing
actuaries. Target-premium tables are included only so that class
actuarial_table can be hidden completely from the GPT and MEC servers
(otherwise, they'd still need that class for targets only).

Achieving that proximate goal also makes it trivial to provide specimen
implementations of old products with a new CSO table.

In the longer term, most or all tables can be consolidated in this new
TU, making it the sole point of access to class actuarial_table.