Make 'AgentPhone' a "required" input field
Some states require the agent's phone number on illustrations.

A validating telephone-number class might be used if one were built into
wxWidgets. There being none, no elaborate validation is warranted. An
end user bent on violating the requirement could enter "(000) 000-0000"
in a validating control as easily as "" in a wxTextCtrl.