Fix an anomaly identified by gcc-8.3
Created an object as 'C {x,y,z}' instead of as 'C({x,y,z})' to avoid an
apparent ambiguity that gcc-8.3 considers an error but gcc-7.2 doesn't.
(Actually, '(C {x,y,z})' is written--the outer parentheses are needed
because the expression is a macro argument--but that has nothing to do
with the C++ issue at hand.)

Paraphrased gcc diagnostics:

In 'static void input_test::test_product_database()'
  /opt/lmi/src/lmi/input_test.cpp:229:86:
with initializer list:
  ({mce_male, mce_rated, mce_smoker, 100, mce_medical, mce_s_XX})

error: call of overloaded
  'database_index(<brace-enclosed initializer list>)' is ambiguous
candidates:
  'database_index::database_index(std::array<int, 6>)'
  'constexpr database_index::database_index(const database_index&)'
  'constexpr database_index::database_index(database_index&&)'

This report seems similar:
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=84849
and the maintainers seem to have accepted it as a defect, or at least
not to have rejected it as not-a-defect.