Expunge extraordinary variable 'HasComplianceTrackingNumber'
* reg_d_group_narr_summary2.mst: Removed the entire 'imprimatur' block,
including its legend and its conditional.
* pdf_command_wx.cpp: Expunged 'HasComplianceTrackingNumber', its
unusual implementation, and the accompanying citation.

Because of recent changes, 'imprimatur' is now shown (unconditionally)
on the cover page for Reg D illustrations, so it need no longer be shown
(conditionally) on a later narrative page. 'HasComplianceTrackingNumber'
was used only to implement that condition, so it's no longer needed at
all; its removal is especially welcome because it had been harder to
define than other variables, as explained here:
  https://lists.nongnu.org/archive/html/lmi/2019-02/msg00001.html