Refactor census view to enable several implementations
Extract all code specific to wxDataViewCtrl to a new CensusDVCView class,
keeping only the code independent of the control being used by the view
in the base CensusView class, in order to allow adding a wxGrid-based
implementation in the upcoming commits.

CensusDVCView function implementations was not sorted properly to
decrement changes count and be sure that real logic changes was not
made at all.