Always prefer '<br>' to '<p></p>'
Now that empty paragraphs are suppressed, they must not be used to add
white space. In this case, '<br><br><br>' has the same visual effect as
the empty paragraph it has replaced. See:
  https://lists.nongnu.org/archive/html/lmi/2019-07/msg00007.html
for a discussion that quotes:
  https://www.w3.org/TR/REC-html40/struct/text.html#edef-P
| We discourage authors from using empty P elements.
| User agents should ignore empty P elements.

This was the only empty paragraph in '*.mst'. To find all, a command
such as this might be used:
  grep -zo '< *[Pp] *>[[:space:]]*< *\/[Pp] *>' *.mst
if newlines can occur only between '<>'-enclosed tags. It would be
tedious to replace every ' ' in that command with '[[:space:]]', but a
complete search can conveniently enough be done from the command line
using vim:
  vim -p --cmd "vim '<\_s*[Pp]\_s*>\_s*<\_s*\/\_s*[Pp]\_s*>' *.mst"
(Then type ':qa!' to exit vim.)