Move "ultimate" rate from MST header to '.policy'
The convoluted conditionals ported from XSL-FO:
  if SinglePremium
    if ModifiedSinglePremium0
      then XXX
    else
      then YYY
    endif
  endif
are no longer required: instead, each proprietary product has
  "Ultimate Illustrated Crediting Rate: {{AnnGAIntRate_Current[5]}}¶"
or the same with index 10, or a blank string. Thus, the pure mustache
implementation was logic-riddled, but its replacement is more nearly
logic-free.

The residual inforce-year logic could similarly be moved, but a major
release is coming very soon, so a conservative approach is better; and
the logic ported from XSL-FO was flawed anyway:
  if year < 5 then print Nth-year value; else print nothing;
is specified for N ∈ {5,10}, where clearly this is wanted:
  if year < N then print Nth-year value; else print nothing;
            ^