Simplify single-premium conditionals
Replaced four booleans with these two:
 - {{SinglePremium}} (whether "Modified" or not), and
 - {{ModifiedSinglePremium}} ("Modified" only).
Each product has its own '.policy' and '.database' files, which already
embody all its parameters. The old XSL-FO logic attempted to deduce some
parameters by parsing 'PolicyLegalName', but that approach was fragile,
difficult to understand, and unnecessarily complicated: in effect, it
did something like this in an lmi C++ file:
  add_variable("product_deduced_to_be_sample2_naic", complex_code);
and this in an MST file:
  {{#product_deduced_to_be_sample2_naic}}
    Special-case string value.
  {{/product_deduced_to_be_sample2_naic}}
  {{^product_deduced_to_be_sample2_naic}}
    {{SomeStringVariable}}
  {{/product_deduced_to_be_sample2_naic}}
whereas the proper approach is to make the variable vary by product,
e.g., in 'sample2_naic.policy':
  <SomeStringVariable>Special-case string value.</SomeStringVariable>
A product's database is always consulted, and that lookup constitutes
a switch on product already: replicating that switch in MST templates
was a grave mistake.