Use no-lapse footnote already in '.policy' files
Reusing strings already defined in '.policy' files, instead of repeating
their contents in line, obviously makes code more maintainable. It also
helps identify subtle errors--such as here, where the name of one type
of no-lapse guarantee (that is limited to a given age or duration) had
incorrectly been combined with the description of a different guarantee
(that has no such limit), and printed whenever the weaker guarantee was
provided--but nothing was printed if the stronger guarantee was offered.