Fix "Superheader width exceeds available" for indiv PP, naively
Now double-digit interest rates can be illustrated. But this fix,
committed only as a demonstration, is not really satisfactory. The real
problem is not that the superheader is too wide for the space allotted,
but rather that it should span four columns instead of two. That is, it
applies not only to the CSV and DB columns, but also to their associated
IRR columns. Therefore, this commit will be overridden in the next
commit, which will provide the proper fix.