Turn narrative chroot instructions into scripts
It's easier to run nine shell scripts than to cut and paste a couple
hundred command lines.

This commit simply moves command lines, to make review easier. It is
probably possible to write one script to invoke the rest in order, on
the host system or in the chroot as appropriate. It should be simple
enough to write a different script that invokes some of the rest as
required to set up a dedicated lmi server that doesn't use a chroot.

Not all of these need to be zsh scripts: many can be POSIX, with
various necessary changes that may be made later.