Reduce obscurity
This "include" file is to be sourced, so it needn't 'export' the
variables it sets; in particular, it shouldn't export only one of
several variables. The original idea was to avoid SC2034: it seemed
at first that writing 'export' was less obtrusive and unnatural than
adding a shellcheck directive; but exporting only one of three
variables now seems more unnatural, and the tersest comment capable
of explaining it is the shellcheck directive (which alone was not
previously sufficient because it was misplaced).