Use long options for 'xargs' throughout lmi
Although it is "deprecated", '--replace' is preferable to '-I'. An '-I'
option is always confusing, because it looks like either '-l' or '-1' in
too many fonts. '--replace', meaning '-i', is deprecated by POSIX only
because "-i" has an implicit "'{}'" argument; but "-i '{}'" and thus
"--replace='{}'" are perfectly okay. If the GNU project had defined a
long option for '-I', that would be perfect.

Incidentally replaced '-P 0' by '--max-procs="$(nproc)"', because:
  $xargs --show-limits
  Maximum parallelism (--max-procs must be no greater): 2147483647
and, given enough arguments, '-P 2147483647' might do something insane.
The imperfect portability of 'nproc' is preferable to insanity.