Work around an apparent 'schroot' anomaly
If the values of both 'users' and 'groups' are double-quoted in
  /etc/schroot/chroot.d/*.conf
e.g.
  users="greg"
  groups="greg"
then the specified normal user cannot enter the chroot:

$schroot --chroot=centos7lmi --user=greg
E: Access not authorised
I: You do not have permission to access the schroot service.
I: This failure will be reported.

Prefixing that failing command with 'sudo' makes it work as expected.

This might look like a problem with
 (1) the setuid bit on /usr/bin/schroot, or
 (2) permissions on the chroot's root directory (0700, e.g.),
but the same user can enter other chroots, which rules out (1), and
(2) is ruled out by inspection:
  $ls -ld /srv/chroot/centos7lmi
  drwxr-xr-x 17 root root 12288 Oct  9 15:56 /srv/chroot/centos7lmi