Fix wrong use of AC_CHECK_PROG() in configure
The variable containing the path to xmlwrapp-config was erroneously set
to "yes", instead of the full path to the config script, preventing this
from working (so that XMLWRAPP_CFLAGS and XMLWRAPP_LIBS always had to be
specified explicitly for MSW cross-builds before).

Fix this by using less error-prone AC_CHECK_PROGS() macro version.

Also rename "no" fallback value to "not-found" for clarity.