Move a PDF paragraph
'SinglePremiumFootnote', if nonempty, has the form of a definition,
e.g.:
  Single Premium: This term means blah blah blah...
In some circumstances, the term "Modified Single Premium" is used
instead, and a "Limited Single Premium" alternative will soon be
needed: the term changes, but the definition remains the same.

This definition is part of a list that has hitherto been alphabetized
strictly, causing its position in the list to shift depending on the
term's first letter. Now that precious refinement is being eliminated
rather than letting the new "Limited" case make the code even more
complicated. This commit merely rearranges the definitions, so that
'SinglePremiumFootnote' is alphabetized as though it always began with
the word "Single".