Explicitly test dubious assumptions underlying some unit tests
The code covered by this unit test is actually defective, and
appropriately misbehaves with clang. That defect is masked by a
defective assumption in the unit test. Result: the code "works" with
gcc, and the unit test appears to demonstrate that it's correct, even
though the code, the test, and gcc are all actually defective. See:
  https://lists.nongnu.org/archive/html/lmi/2020-01/msg00011.html