Improve (s)chroot detection
There is no simple and portable way to determine whether a script is
running in a chroot. Outside a chroot, 'ls -di /' prints "2" for ext4,
but "64" for a particular server with an XFS filesystem using LVM; it's
unknown what it might print in general, but "32" is also reported.

Because the 'lmi_setup*.sh' scripts use only schroot, it's better to
detect schroot directly, by looking for a variable it always defines
and assuming that no one is so irrational as to define it outside a
chroot environment.