Detect whether gnu.org is, bizarrely, blocked
It is better to test dynamically whether gnu.org is reachable than to
embed that information in a configuration file that must be edited.
'curl' seems to be the best tool for this, and it deserves to be made
generally available. It might even be handy for cygwin, though there
is no need to update a cygwin installation now just to get 'curl', and
it is hoped that cygwin support can be withdrawn in favor of either
"WSL" or GNU/Linux.