Specify UTC in case TZ is otherwise
A corporate redhat server seems to use a local timezone. It seems
sufficient to change only this redhat-specific script, because all
other work is done in a debian chroot. Probably this change does
nothing in practice, unless the script is run almost at the moment
of a daylight-saving change, but it's simplest to make that problem
impossible by specifying '-u' consistently.