Streamline temporary instrumentation
Don't measure the time it takes to authenticate using an external
binary, because internal routines are always preferable. For now,
use first one of these commands, then the other
  wine ./lmi_wx_shared --data_path=/opt/lmi/data
  wine ./lmi_wx_shared --data_path=/opt/lmi/data --pyx=measure_md5
to run some scenario like
  File | New | Illustration
  OK
  File | Print to PDF
and compare the elapsed time shown on the statusbar, to see the cost
of reauthenticating before generating each PDF.

On my machine, running under 'wine', I see:
  341 msec without '--pyx=measure_md5'
  405 msec  with   '--pyx=measure_md5'
and (405-341)/341 is about a twenty-percent penalty.