Augment 'show_env' makefile target, exposing a defect
Running
  make check_concinnity
in a new terminal fails. Running
  make show_env
shows one symptom--even though these variables are set as intended:
  LMI_COMPILER    = "gcc"
  LMI_TRIPLET     = "i686-w64-mingw32"
this one
  exec_prefix     = "/opt/lmi/_"
doesn't have the value expected from its definition:
  exec_prefix     := $(prefix)/$(LMI_COMPILER)_$(LMI_TRIPLET)
because the $(LMI_COMPILER) and $(LMI_TRIPLET) definitions are not in
effect when $(exec_prefix) is set. Recursive expansion, i.e.:
-  exec_prefix     := $(prefix)/$(LMI_COMPILER)_$(LMI_TRIPLET)
+  exec_prefix     = $(prefix)/$(LMI_COMPILER)_$(LMI_TRIPLET)
would fix that symptom, but wouldn't make 'check_concinnity' work.

Added $(EXEEXT) and $(SHREXT) to the list of environmental variables
displayed, so that
  make show_env
shows:
  EXEEXT          = ""
  SHREXT          = ".so"
demonstrating that $(LMI_COMPILER) and $(LMI_TRIPLET) are also not set
when 'configuration.make' is included.

Running
  . /opt/lmi/src/lmi/set_toolchain.sh
whenever a new terminal is started works around this problem, but it is
better to resolve the defect.