Eradicate bogus complexity in a footnote
The NAIC illustration reg, (7)(B)(2), says:

"For a policy that does not require payment of a specific contract
premium, the illustration shall show the premium outlay that must
be paid to guarantee coverage for the term of the contract, subject
to maximum premiums allowable to qualify as a life insurance policy
under the applicable provisions of the Internal Revenue Code"

That guaranteed premium may be zero. In that case: formerly, this
footnote said "a [modal] premium is defined", which fails to "show
the premium outlay that must be paid" as the regulation requires.

In this test case:
  File | New | Illustration
  "Policy": sample2_naic
  "Effective date": one year before today
  "General account value": 500000
  OK
this guaranteed premium is indeed zero. For this million-dollar policy
with a half-million-dollar account value, coverage is guaranteed for
the full term of the contract, even if no further premium is paid.