Improve physical design
Instead of writing stream inserters and extractors as 'inline' functions
in a header (which must then include <[io]stream>), only declare them in
the header (where <iosfwd> suffices), and implement them in a '.cpp'
file (which includes <[io]stream>).

Don't do this for header-only facilities.

Don't touch this inserter in 'path_utility.hpp':
  inline std::ostream& operator<<(std::ostream& os, fs::path const& z)
because the standard <filesystem> header provides such an inserter. It
would not be good to add a 'path_utility.o' dependency now, only to
remove it when <filesystem> becomes usable.

Incidentally, avoid dummy argument names in the header, such as:
  std::istream& operator>>(std::istream& TheStream, foo& TheFooThing);
They're just clutter, and over years of maintenance they can easily come
to differ from the argument names used in the implementation.