Nychthemerally build c[gl]i tests with shared-object attributes
This line in 'GNUmakefile':
  $(MAKE) all cgi_tests cli_tests build_type=so_test USE_SO_ATTRIBUTES=1
suggests that such a build succeeds; and whatever can be built, should
be built nychthemerally.

Incidentally renamed log file for shared-object build of 'all:' target.
The corresponding c[gl]i log added by this commit is named:
  "$log_dir"/cgi_cli_so_test
so the original log name for 'all:':
  "$log_dir"/so_test
is too broad, and prepending the name of the target:
  "$log_dir"/all_so_test
seems even more emphatically overbroad, so the unwieldy but
non-misleading name:
  "$log_dir"/default_targets_so_test
seems to be the least bad idea.