Regularize compositing of non-additive vectors
Formerly, these vectors had potentially different lengths than the
additive vectors in a composite, and most were all zero. Now, at
least they all have the appropriate length, so that accessing the
value in the year preceding maturity won't cause a crash. The value
for each (yearly) element is overwritten by the last addend that
contains such an element; that's what 'LedgerVariant' does for such
vectors already, and it's arguably the best simple, uniform method
that can be devised.