Simplify [348]
Almost all strings are composited in the same way. Accordingly, saved
a couple hundred lines of code by handling the whole collection
generically without naming its members, and writing special-case code
only for the few exceptions to the general rule.

Incidentally demoted a marked defect. A plain comment suffices to point
out that some strings perhaps shouldn't be allowed to vary by life. But
it is difficult to name one that could never vary--even ContractNumber
might be the same, if it is blank, or if a census is used to depict the
same policy under various scenarios.