Regularize mounting
Several cache directories are mounted in chroots, for guest-OS packages
as well as for lmi. Regularized commands to create directories, using
'mkdir -p' in all cases, for robustness: for example,
  /var/cache/"${CODENAME}"
had been tacitly assumed to exist, but that assumption is false the
first time the script is run in a pristine environment; instead, that
directory now is explicitly created with 'mkdir -p' to prevent the mount
command from failing. Added commands to show the disk usage of each
mountpoint before mounting: if it isn't pretty much empty, then there
may be something wrong. Where mount commands were previously allowed to
fail (causing cachable files to be written to guest directories only),
now they are prevented from failing (creating host directories for files
that can profitably be cached and shared).