Create apt cache directory on host before running debootstrap
Previously, the files that debootstrap downloaded were saved in a
directory that was later obscured by a mount, so that they weren't
preserved across reinstallations (as demonstrated by running 'du'
before mounting). Now, downloads are cached on the host (i.e., the
ultimate debian-stable host for a
  debian-testing within centos within debian-stable
chroot chain). This seems to save a minute or two on each post-
initial run of 'install_centos.sh'. This partial diff of logs
before (-) and after (+) this change...

 debootstrap "bullseye" /srv/chroot/"lmi_bullseye_1" http:[...]
[...]
-I: Retrieving libacl1 2.2.53-6
-I: Validating libacl1 2.2.53-6
+I: Validating libacl1 2.2.53-8
[...the newer libacl1 had already been retrieved...]
-I: Retrieving adduser 3.118
 I: Validating adduser 3.118
-I: Retrieving libapparmor1 2.13.4-1+b1
 I: Validating libapparmor1 2.13.4-1+b1
[...snip one hundred fifty-one others...]
-I: Retrieving liblzma5 5.2.4-1+b1
 I: Validating liblzma5 5.2.4-1+b1
-I: Retrieving zlib1g 1:1.2.11.dfsg-2
 I: Validating zlib1g 1:1.2.11.dfsg-2

...demonstrates that packages previously retrieved are now taken
from cache rather than retrieved again.