Stop imposing the use of C++17 in configure
It's possible (or will soon become possible) to use C++20 too, so don't
override the CXX value specified by the user from configure itself.

Note that this still preserves the check for C++17 support, so if the
compiler doesn't use C++17 or later by default and -std=c++17 or c++20
wasn't specified in CXX, configure will now fail. This is not especially
user-friendly, but more flexible.