Replace DatumSequenceEditorEvtHandler with a lambda
There is no need to define a class just in order to bind a single event
handler, this can be done much more easily and concisely by using Bind()
with a lambda.

The only non-obvious part is that we have to call Bind() on the extra
event handler used for the grid editors and not the control itself, so
add a comment explaining it.