Remove deleted DatumSequenceEditor copy-related methods
There is no need to delete them explicitly, as they're already deleted
in the base wxGridCellEditor class. Besides, doing it only in this class
but not in the other specialized grid editors was inconsistent and it
seems better to stop doing it here rather than also doing it elsewhere.

No real changes.