Replace table_type_converter::get_all() with register_all()
We don't really need to have a map with all the converters at all, all
we need is a way to register them and this can be done with a dedicated
function.

And this function can be implemented by simply registering all the
custom converters, there is no any need to create a map just to iterate
over it.