Remove separate table_type_converter::type()
The difference between it and grid_value_type() was not obvious and
only the latter one is actually needed, so remove type() from the base
class and replace it with value-independent grid_type() for the custom
type converters for which it can be implemented generically, in terms of
typeid().

As table_custom_type_converter is now a template class, register_all()
can't be its method any longer, so just inline it at the point of use.