Remove two failing commands
Resolved the issues left open by commit 294b7bb24 of 20200521T2317Z:

| Let certain commands fail for the moment

Removed those commands.

| Dot files cannot be set for a user that does not yet exist.

Therefore, those commands always failed.

| Probably 'lmi_setup_21.sh' already takes care of this

Indeed it does set up .zshenv and .vimrc in the innermost chroot.
But it does not do so for the intermediate centos chroot; that
need will be addressed in the next commit.