Remove unwanted files before copying them
The first time a system test is performed in a new chroot, it finds no
touchstone to compare against, so a spurious error report is issued.
That behavior is correct in general, but is disregarded for new chroots
(which presumably use only previously-accepted code).

Previously, freshly-generated results were copied to the touchstone/
directory, and then datestamped analyses were removed therefrom. Now,
datestamped analyses are first removed, then only the remaining files
are copied. This may seem to be a needless refinement, but it has the
very desirable effect of causing the expanded copy command to be
identical between runs that use the same testdecks, so that successive
logs don't routinely mismatch on a 50000-character line, starting from
byte 7000 or so.

Removing datestamped analyses from the test/ directory does no harm:
because the first system test failed completely, they contain no useful
information.