Use 'wget -N' even for servers that don't offer time stamps
With github.com, 'wget -N' elicits this message:
  Last-modified header missing -- time-stamps turned off.
but '-N' still has an important effect: it causes any existing copy of
the same file to be overwritten. (That's the behavior described in the
gnu wget manual for a different set of circumstances:
  | When running Wget with ‘-r’ or ‘-p’, but without ‘-N’, ‘-nd’,
  | or ‘-nc’, re-downloading a file will result in the new copy
  | simply overwriting the old.
so, in effect, an unavailable remote timestamp is treated as older than
any local copy that already exists.)

Thus, for servers that don't provide timestamps, wget without '-N'
doesn't overwrite an existing copy; instead, it saves a fresh copy with
a serial suffix (".1", ".2", etc.), which is clearly not desirable here.