Handle a new warning differently
Commit c9d97d506bd of 20200428T2154Z added '-Wno-deprecated-copy' to
g++ flags in selected cases, to suppress a warning added in gcc-9,
although gcc-8 is still used in production. Unrecognized '-Wno-'
options are silently ignored...except when they are not:

  https://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html

| When an unrecognized warning option is requested (e.g.,
| -Wunknown-warning), GCC emits a diagnostic stating that the
| option is not recognized. However, if the -Wno- form is used,
| the behavior is slightly different: no diagnostic is produced
| for -Wno-unknown-warning unless other diagnostics are being
| produced. This allows the use of new -Wno- options with old
| compilers, but if something goes wrong, the compiler warns
| that an unrecognized option is present.

The "if something goes wrong" effect being a nuisance, removed the
'-Wno-deprecated-copy' flag from makefiles, which as of the last
commit use '-isystem' to suppress this diagnostic in third-party
code. If it is caused by lmi code, that code should be fixed.

There is no need to remove '-Wno-deprecated-copy' from C++ files where
it is readily restricted to particular gcc versions. Such a restriction
cannot reasonably be coded in a makefile.

Tested with gcc-8.3 only.